Tema3

+++ src/random/Generator.java
@@ -2,5 +2,5 @@ package random;
 
 public interface Generator {
 
-	double next();
+	float next();
 }
